Move more stuff to its own file.
